Added lots of functionality to do layout in JavaScript. This is really slow, but it gets closer to the layout I am looking for. Hard-coding some more sizes may be an option that would allow for getting rid of some JavaScript layout.
The new layout fixes issues with bin numbers getting pushed off the side of the column. It also fixes issues with varietal and type headings getting orphaned from their first child.

Now using non-breaking spaces only in the parts of the locale that are not separated by commas instead of using nowrap for the entire locale.

Fixed UTF-8 encoding issue that showed up in Firefox (new meta tag in the page head).