Removed unused wrap elelement functions (that functionality was moved to XSLT in an earlier commit).
Switched to a layout that floats both the wine name and the bin number. This way the height of each wine element doesn't need to be updated.

Implemented a new updateLayout function that computes a mininum width for the entire wine list so that names and bin numbers don't overlap.